pyproject.toml

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poster_maker"
version = "1.0.0"
description = "Split images into equal parts for poster creation"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Sumit", email = "sumit749284@gmail.com"}
]
dependencies = [
    "Pillow>=9.0.0",
    "PyYAML>=6.0",
    "tqdm>=4.62.0",
    "psutil>=5.8.0",
    "reportlab>=3.6.0",
    "questionary>=2.0.0"
]

[project.scripts]
poster-maker = "poster_maker.main:main"

[tool.setuptools]
packages = ["poster_maker"]

[project.urls]
"Homepage" = "https://github.com/smaxiso/poster_maker"
"Bug Tracker" = "https://github.com/smaxiso/poster_maker/issues"